added support for in,out files
run iseq as

 ./iseq inputfile outputfile

input file is optional and so is output file, but if outputfile is
specified then input file must be specified as well. if input file is
"-" then input is from stdin; if output file is "-" then output is to
stdout.

The default, when no files are specified, is to use stdin and stdout.